Move permissions system to another assembly
