+++ Assets/Scripts/SS3D/Permissions/PermissionSystem.cs.meta
