+++ Assets/Scripts/SS3D/Permissions/SS3D.Permissions.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8979f103c1228714eb0115e440b2b33a
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
