+++ Assets/Scripts/SS3D/Systems/Health/BodyParts/Bodypart.cs
@@ -9,7 +9,6 @@ using SS3D.Systems.Health;
 using System.Linq;
 using System.Collections.ObjectModel;
 using FishNet;
-using Coimbra;
 using SS3D.Systems.Inventory.Containers;
 using SS3D.Systems.Inventory.Items;
 using System;
