+++ Assets/Scripts/SS3D/Systems/IngameConsoleSystem/Commands/AddHandCommand.cs
@@ -1,14 +1,13 @@
 ﻿using SS3D.Core;
 using SS3D.Systems.Entities;
-using SS3D.Systems.Permissions;
 using SS3D.Systems.PlayerControl;
 using UnityEngine;
 using FishNet;
 using SS3D.Data;
 using SS3D.Data.Enums;
-using FishNet.Object;
 using SS3D.Systems.Inventory.Containers;
 using FishNet.Connection;
+using SS3D.Permissions;
 
 namespace SS3D.Systems.IngameConsoleSystem.Commands
 {
