+++ Assets/Scripts/SS3D/Systems/IngameConsoleSystem/Commands/Command.cs
@@ -1,5 +1,5 @@
 ﻿using FishNet.Connection;
-using SS3D.Systems.Permissions;
+using SS3D.Permissions;
 
 namespace SS3D.Systems.IngameConsoleSystem.Commands
 {
