+++ Assets/Scripts/SS3D/Systems/IngameConsoleSystem/Commands/EchoCommand.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Linq;
 using FishNet.Connection;
-using SS3D.Systems.Permissions;
+using SS3D.Permissions;
 
 namespace SS3D.Systems.IngameConsoleSystem.Commands
 {
