+++ Assets/Scripts/SS3D/Systems/IngameConsoleSystem/Commands/ItemCommands/DescribeItemCommand.cs
@@ -1,8 +1,8 @@
 ﻿using FishNet.Connection;
+using SS3D.Permissions;
 using SS3D.Systems.Inventory.Items;
-using SS3D.Systems.Permissions;
 
-namespace SS3D.Systems.IngameConsoleSystem.Commands
+namespace SS3D.Systems.IngameConsoleSystem.Commands.ItemCommands
 {
     /// <summary>
     /// Command to describe an item held in the active hand of the player calling the command.
