+++ Assets/Scripts/SS3D/Systems/IngameConsoleSystem/Commands/PlayerListCommand.cs
@@ -2,8 +2,8 @@
 using System.Linq;
 using FishNet.Connection;
 using SS3D.Core;
+using SS3D.Permissions;
 using SS3D.Systems.Entities;
-using SS3D.Systems.Permissions;
 using SS3D.Systems.PlayerControl;
 
 namespace SS3D.Systems.IngameConsoleSystem.Commands
