+++ Assets/Scripts/SS3D/Systems/IngameConsoleSystem/Commands/QuitCommand.cs
@@ -1,6 +1,5 @@
 ﻿using FishNet.Connection;
-using SS3D.Systems.Permissions;
-using UnityEngine;
+using SS3D.Permissions;
 
 namespace SS3D.Systems.IngameConsoleSystem.Commands
 {
