+++ Assets/Scripts/SS3D/Systems/IngameConsoleSystem/Commands/ReconnectCommand.cs
@@ -1,6 +1,6 @@
 ﻿using System.Diagnostics;
 using FishNet.Connection;
-using SS3D.Systems.Permissions;
+using SS3D.Permissions;
 using UnityEngine.Device;
 
 namespace SS3D.Systems.IngameConsoleSystem.Commands
