+++ Assets/Scripts/SS3D/Systems/IngameConsoleSystem/CommandsController.cs
@@ -6,8 +6,8 @@ using FishNet.Connection;
 using FishNet.Object;
 using SS3D.Core;
 using SS3D.Core.Behaviours;
+using SS3D.Permissions;
 using SS3D.Systems.IngameConsoleSystem.Commands;
-using SS3D.Systems.Permissions;
 using SS3D.Systems.PlayerControl;
 using UnityEngine;
