+++ Assets/Scripts/SS3D/Systems/PlayerControl/PlayerSystem.cs
@@ -7,8 +7,8 @@ using FishNet.Transporting;
 using SS3D.Core;
 using SS3D.Core.Behaviours;
 using SS3D.Logging;
+using SS3D.Permissions;
 using SS3D.Systems.Entities;
-using SS3D.Systems.Permissions;
 using SS3D.Systems.PlayerControl.Events;
 using SS3D.Systems.PlayerControl.Messages;
 using UnityEngine;
