+++ Assets/Scripts/SS3D/Systems/Rounds/RoundSystemBase.cs
@@ -7,7 +7,7 @@ using FishNet.Object.Synchronizing;
 using SS3D.Core;
 using SS3D.Core.Behaviours;
 using SS3D.Logging;
-using SS3D.Systems.Permissions;
+using SS3D.Permissions;
 using SS3D.Systems.PlayerControl;
 using SS3D.Systems.Rounds.Events;
 using SS3D.Systems.Rounds.Messages;
