+++ Assets/Scripts/SS3D/Systems/SS3D.Systems.asmdef
@@ -19,7 +19,8 @@
         "GUID:9e24947de15b9834991c9d8411ea37cf",
         "GUID:75469ad4d38634e559750d17036d5f7c",
         "GUID:eb124556ba20d60469be9771eb1f0360",
-        "GUID:46e20dca70571ce4e817a6b5874168a0"
+        "GUID:46e20dca70571ce4e817a6b5874168a0",
+        "GUID:8979f103c1228714eb0115e440b2b33a"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
